contextmenu and deployment
