+++ packages/builder/src/stores/builder/deployments.ts
@@ -1,11 +1,12 @@
-import { writable } from "svelte/store"
+import { writable, type Writable } from "svelte/store"
 import { API } from "api"
 import { notifications } from "@budibase/bbui"
+import { DeploymentProgressResponse } from "@budibase/types"
 
 export const createDeploymentStore = () => {
-  let store = writable([])
+  let store: Writable<DeploymentProgressResponse[]> = writable([])
 
-  const load = async () => {
+  const load = async (): Promise<void> => {
     try {
       store.set(await API.getAppDeployments())
     } catch (err) {
